pyproject.toml
[project]
name = "law-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gradio>=6.9.0",
    "olefile>=0.47",
    "openai>=2.26.0",
    "openpyxl>=3.1.5",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
